refactor(loader): 合并 satellite-aware loader 到 packagesload，OBS-01 覆盖 satellite [#2147]
GREEN：把 typeseval 与 metricschema 两套 loader 的 satellite 前缀展开 + member
分组下沉到唯一共享入口 packagesload.LoadWorkspace；OBS-01 metric-PII scan 经此
覆盖 cmd/adapters/examples satellite 生产代码。

- packagesload.LoadWorkspace（新）：唯一 satellite-aware loader，吸收 typeseval 的
  workspacePatternGroups/splitWorkspacePattern/matchWorkspaceMember/loadPackageGroups
  + metricschema 的 per-path dedup。satellite 展开原子 ExpandParentPrefix 从
  tools/workspace 迁入设为私有 expandParentPrefix（包外不可表达）——配合既有
  PACKAGES-LOAD-FUNNEL-01，「别处重造 satellite loader」不可表达（单 loader 不变式 Hard 化）。
- typeseval.loadPackagesMode delegate LoadWorkspace；LoadPackages/SharedResolver 签名不动。
- metricschema.loadPackages delegate LoadWorkspace + packageHasProjectFile 过滤；
  obs01ProductionPatterns = Patterns + prodscan.SatelliteParentPatterns（单源增量，
  不经 PatternsExtended 拖入 tests/tools）。Build/loadReachablePackages 走语义不同的
  跨模块可达路径，不并入（carve-out）。
- workspace 新增 HasGoWork，删除导出的 ExpandParentPrefix。
- prodscan 抽 SatelliteParentPatterns 单源增量；重写两-loader 拆分 godoc（已合并为单 loader）。
- 测试：packagesload 表测 expandParentPrefix/splitWorkspacePattern + go.work 集成；
  splitWorkspacePattern 测试随函数迁入；satellite RED fixture/RequiresSatellites 转 GREEN。

验证：tools 全 build；metricschema 全套 + packagesload + prodscan + workspace + typeseval 绿；
全量 archtest 套件 vs origin/develop 失败集逐一相同（17 个均 pre-existing，零回归）；
真实仓库 OBS-01（TestMetricLabelErrcodeClassifiersRequireAck）绿——cmd/adapters/examples
无 errcode-classifier→metric-label 泄露 sink；golangci-lint 0 issues。
