+++ tools/internal/prodscan/patterns.go
@@ -13,20 +13,20 @@ import (
 //   - a single-module consumer/fixture project (the metricschema OBS-01 fixtures,
 //     a downstream module that vendors gocell).
 //
-// dirExists + IsModuleRoot prune entries that don't apply, and patternLoadMode
-// (in tools/metricschema) routes each surviving pattern to the right load mode, so
-// the EFFECTIVE scan differs per context without a per-context list:
+// dirExists + IsModuleRoot prune entries that don't apply, and the shared
+// satellite-aware loader (packagesload.LoadWorkspace) routes each surviving pattern
+// to the right load mode — expanding a multi-member parent prefix to its members —
+// so the EFFECTIVE scan differs per context without a per-context list:
 //
 //   - Real repo post-#1565: the core lives in the `framework` submodule, so
 //     framework/{kernel,runtime,pkg} match (loaded ModeWorkspace — framework is a
 //     sibling go.work member); the bare kernel/runtime/pkg no longer exist at the
 //     root and are pruned; cmd/adapters/examples each hold SEVERAL go.work member
 //     modules (multi-member parents), so their "./<dir>/…" pattern is owned by no
-//     single member and is pruned by HasNestedModuleRoot from THIS satellite-free
-//     scan (metricschema's OBS-01 loader cannot expand a multi-member prefix). The
-//     typeseval-backed duration gates opt INTO satellite coverage via
-//     PatternsWithSatellites; OBS-01 metric-PII coverage of satellites stays gh
-//     #2136. cellmodules/corecells are module roots → pruned by IsModuleRoot.
+//     single member and is pruned by HasNestedModuleRoot from THIS base scan — it is
+//     re-emitted ONLY by SatelliteParentPatterns, which callers compose in when they
+//     want satellite coverage (the shared loader expands it to the members).
+//     cellmodules/corecells are module roots → pruned by IsModuleRoot.
 //   - Single-module fixture: framework/<layer> doesn't exist (pruned); the bare
 //     cmd/pkg/kernel/runtime the fixture writes ARE part of the one module, so their
 //     "./<dir>/…" pattern matches under ModeModule and is actually scanned.
@@ -48,7 +48,7 @@ var topLevelDirs = []string{
 // governance scanners. Missing top-level directories are skipped so temp-module
 // fixtures can exercise the production entrypoints without synthetic dirs.
 //
-// Two kinds of top-level dir are pruned from this satellite-FREE scan:
+// Two kinds of top-level dir are pruned from this satellite-FREE base scan:
 //
 //   - A go.work satellite MODULE ROOT (carries its own go.mod — e.g. cellmodules/
 //     after #1559, corecells/): a single module addressed by its own member
@@ -56,27 +56,17 @@ var topLevelDirs = []string{
 //   - A MULTI-MEMBER PARENT (cmd/, adapters/, examples/): not a module root itself,
 //     but holding several go.work member modules one level deeper (cmd/gocell +
 //     cmd/corebundle; adapters/postgres …; examples/<id> …). "./cmd/..." is owned by
-//     no single member, so post-#1565 (no root module to anchor it) it would
-//     hard-error or silently match-zero (HasNestedModuleRoot).
+//     no single member (HasNestedModuleRoot).
 //
-// Both are pruned so this PRODUCTION scan stays loadable under metricschema's OBS-01
-// loader, which loads each pattern in ModeModule/ModeWorkspace and does NOT expand a
-// multi-member parent prefix. A single-module fixture writes cmd/pkg/... as part of
-// its ONE module (no nested go.mod), so neither predicate fires there and the
-// fixture's "./cmd/..." is kept and scanned.
-//
-// Rules that DO require satellite coverage opt in through the typed (typeseval)
-// loader, which expands a satellite parent prefix to its real go.work members via
-// workspace.ExpandParentPrefix (kept honest by SATELLITE-PARENT-PREFIX-SCAN-01):
-//
-//   - the broad duration gates (TEST-TIME-LITERAL-01, PROD-DURATION-CONST-01) take
-//     the whole satellite set via PatternsWithSatellites; and
-//   - narrow security/authz funnels (fence-token mint, …) hardcode the specific
-//     satellite patterns they need.
-//
-// Both opt-in paths are independent of THIS list — Patterns deliberately does NOT
-// emit the satellite prefixes, since OBS-01's loader cannot expand them (its own
-// metric-PII coverage of satellites remains gh #2136).
+// Patterns is the conservative BASE: it omits the satellite prefixes so consumers
+// that do not want satellite scope (the bulk of the production scanners) are not
+// silently widened. Widening to satellites is a per-gate opt-in: a caller composes
+// SatelliteParentPatterns onto Patterns (OBS-01) or PatternsExtended
+// (PatternsWithSatellites), and the shared satellite-aware loader
+// (packagesload.LoadWorkspace) expands each parent prefix to its real go.work
+// members before loading. A single-module fixture writes cmd/pkg/... as part of its
+// ONE module (no nested go.mod), so neither predicate fires there, the increment is
+// empty, and the fixture's "./cmd/..." stays in Patterns and is scanned directly.
 func Patterns(root string) []string {
 	var patterns []string
 	if dirHasNonTestGoFiles(root) {
@@ -125,47 +115,24 @@ func PatternsExtended(root string) []string {
 	return base
 }
 
-// PatternsWithSatellites widens PatternsExtended(root) with the multi-member
-// satellite parent prefixes ("./cmd/...", "./adapters/...", "./examples/...") that
-// Patterns deliberately prunes.
+// SatelliteParentPatterns returns the multi-member satellite parent prefixes
+// ("./cmd/...", "./adapters/...", "./examples/...") that Patterns prunes — the
+// single-sourced satellite increment. It is DERIVED from the same
+// HasNestedModuleRoot predicate Patterns prunes on (not a hand-maintained list), so
+// a future fourth multi-member parent is picked up automatically; it re-includes
+// ONLY multi-member parents (plain module-root dirs cellmodules/, corecells/ stay
+// excluded by !IsModuleRoot — they are addressed by their own member pattern, not an
+// expandable parent prefix). A single-module fixture has no nested go.mod, so the
+// increment is EMPTY there.
 //
-// USE FOR typeseval-backed gates ONLY. The emitted satellite prefixes are owned by
-// no single go.work member and can only be loaded by a loader that expands them via
-// workspace.ExpandParentPrefix (tools/archtest's typed façade). Do NOT pass the
-// result to a ModeModule/ModeWorkspace loader that does not expand (e.g. metricschema
-// OBS-01's loadPackages) — it would hard-error or match-zero. Those callers use the
-// satellite-free Patterns / PatternsExtended instead.
-//
-// It is the production-scan source-of-record for the typed (typeseval) duration
-// gates — TEST-TIME-LITERAL-01 and PROD-DURATION-CONST-01 — whose loader's
-// workspace.ExpandParentPrefix expands each satellite parent prefix to its real
-// go.work members before loading (kept honest by SATELLITE-PARENT-PREFIX-SCAN-01).
-// Only that path can actually scan the satellites, so only this constructor emits
-// their prefixes (#2136).
-//
-// Two scan constructors exist, split by LOADER CAPABILITY — not by accident — and
-// they must NOT be merged (doing so re-breaks metricschema):
-//
-//   - Patterns / PatternsExtended (satellite-FREE): consumed by metricschema's
-//     OBS-01 loader, which does NOT expand a multi-member parent prefix; handing it
-//     "./cmd/..." (owned by no single go.work member post-#1565) hard-errors or
-//     match-zeroes, so the satellite parents are pruned there.
-//   - PatternsWithSatellites: consumed only by the typeseval-backed duration gates,
-//     which CAN expand the satellite prefixes.
-//
-// The satellite increment is DERIVED from the same HasNestedModuleRoot predicate
-// Patterns prunes on — not a hand-maintained list — so a future fourth multi-member
-// parent is picked up automatically. It re-includes ONLY multi-member parents:
-// plain module-root dirs (cellmodules/, corecells/) stay excluded by !IsModuleRoot,
-// since they are not expandable parent prefixes (their satellite coverage is gh
-// #2136, a separate concern). A single-module fixture has no nested go.mod, so the
-// increment is empty there and this equals PatternsExtended.
-//
-// Do NOT fold the satellite prefixes into PatternsExtended itself: its other
-// consumers (e.g. errcode_invariants) would then silently gain unreviewed satellite
-// scope. Widening to satellites is a per-gate opt-in by name, not a global default.
-func PatternsWithSatellites(root string) []string {
-	patterns := PatternsExtended(root)
+// The emitted prefixes are owned by no single go.work member and are loadable ONLY
+// through the shared satellite-aware loader packagesload.LoadWorkspace, which expands
+// each to its real members (kept honest by SATELLITE-PARENT-PREFIX-SCAN-01). Both
+// the OBS-01 metric-PII scan (#2147) and the typeseval duration gates compose this
+// increment onto their respective base scope; do NOT hand the prefixes to a loader
+// that does not go through LoadWorkspace.
+func SatelliteParentPatterns(root string) []string {
+	var patterns []string
 	for _, dir := range topLevelDirs {
 		full := filepath.Join(root, dir)
 		if dirExists(full) && !IsModuleRoot(full) && HasNestedModuleRoot(full) {
@@ -175,6 +142,27 @@ func PatternsWithSatellites(root string) []string {
 	return patterns
 }
 
+// PatternsWithSatellites widens PatternsExtended(root) — Patterns plus tests/ and
+// tools/ — with the multi-member satellite parent prefixes (SatelliteParentPatterns).
+// It is the production-scan source-of-record for the typeseval-backed duration gates
+// (TEST-TIME-LITERAL-01, PROD-DURATION-CONST-01), whose broad invariant covers
+// production + test/tool support packages + satellites.
+//
+// Two satellite-bearing scopes exist, split by BASE (not loader capability — since
+// #2147 both load through the same packagesload.LoadWorkspace expander):
+//
+//   - PatternsWithSatellites = PatternsExtended + satellites — the broad duration
+//     gates (production + tests/ + tools/ + satellites).
+//   - OBS-01 = Patterns + satellites — production + satellites, WITHOUT tests/+tools/
+//     (composed in tools/metricschema; OBS-01 never scanned tests/ or tools/).
+//
+// Do NOT fold the satellite prefixes into PatternsExtended itself: its other
+// consumers (e.g. errcode_invariants) would then silently gain unreviewed satellite
+// scope. Widening to satellites is a per-gate opt-in by name, not a global default.
+func PatternsWithSatellites(root string) []string {
+	return append(PatternsExtended(root), SatelliteParentPatterns(root)...)
+}
+
 func dirExists(path string) bool {
 	info, err := os.Stat(path)
 	return err == nil && info.IsDir()
@@ -182,10 +170,10 @@ func dirExists(path string) bool {
 
 // IsModuleRoot reports whether dir carries its own go.mod (a go.work satellite
 // module root). Such dirs are unaddressable by root-relative patterns under
-// ModeModule (GOWORK=off) and are skipped by Patterns / PatternsExtended — see
-// the Patterns godoc + gh #2136. Exported so coverage guards that must stay
-// symmetric with the scan set (e.g. metricschema's productionGoTopLevels) share
-// this single source of truth rather than re-deriving the go.mod check.
+// ModeModule (GOWORK=off) and are skipped by Patterns / PatternsExtended — see the
+// Patterns godoc. Exported so coverage guards that must stay symmetric with the scan
+// set (e.g. metricschema's productionGoTopLevels) share this single source of truth
+// rather than re-deriving the go.mod check.
 func IsModuleRoot(dir string) bool {
 	info, err := os.Stat(filepath.Join(dir, "go.mod"))
 	return err == nil && !info.IsDir()
@@ -195,13 +183,15 @@ func IsModuleRoot(dir string) bool {
 // itself a module root, but at least one immediate subdirectory carries a go.mod
 // (a go.work member). cmd/ (over cmd/gocell, cmd/corebundle), adapters/ (over
 // adapters/postgres …), and examples/ (over examples/<id> …) are the post-#1565
-// instances. A root-relative "./<dir>/..." pattern over such a parent is owned by
-// no single member and cannot be loaded as a ModeModule pattern (no root module to
-// anchor it); its production-scan coverage is deferred to gh #2136. Patterns skips
-// these dirs and the OBS-01 coverage guard (metricschema's productionGoTopLevels)
-// must skip them symmetrically — exported so both share this single source rather
-// than re-deriving the nested-go.mod check. A single-module fixture has no nested
-// go.mod under cmd/pkg/…, so this is false there and the fixture dir is kept.
+// instances. A root-relative "./<dir>/..." pattern over such a parent is owned by no
+// single member and cannot be loaded as a plain ModeModule pattern; Patterns prunes
+// it from the base scope, and SatelliteParentPatterns re-emits it for callers that
+// opt into satellite coverage, where packagesload.LoadWorkspace expands it to the
+// members (#2147). Exported so the satellite increment and the OBS-01 coverage guard
+// (metricschema's productionGoTopLevels, which now REQUIRES these parents) share this
+// single source rather than re-deriving the nested-go.mod check. A single-module
+// fixture has no nested go.mod under cmd/pkg/…, so this is false there and the fixture
+// dir is kept in Patterns directly.
 func HasNestedModuleRoot(dir string) bool {
 	entries, err := os.ReadDir(dir)
 	if err != nil {
